check-prompt: handle pre-existing input correctly for prompt type
Unsent input on the line resulted in 'unknown prompt type, since
check-prompt tests for additional buffer text beyond the prompt match.
Now we see whether the final output text provided by the process
filter matches, and if so, assume it is pre-existing input on the
line.  Note that this test could fail if output is split within the
prompt text itself, resulting in a spurious 'unknown type.  This is
deemed to be rare.